chore: move alerts status to cloud
